Change unit tests and remove depricated flag
ACCS1-8
