+++ extension/test/unit/config/config.spec.js
@@ -458,136 +458,6 @@ describe('::config::', () => {
     )
   })
 
-  describe('addCommercetoolsLineItems', () => {
-    it(
-      'when addCommercetoolsLineItems is set as boolean false in config.js, ' +
-        'it should load as false value in module config',
-      async () => {
-        process.env.ADYEN_INTEGRATION_CONFIG = JSON.stringify({
-          commercetools: {
-            ctpProjectKey1: {
-              clientId: 'clientId',
-              clientSecret: 'clientSecret',
-              apiUrl: 'host',
-              authUrl: 'authUrl',
-              authentication: {
-                scheme: 'basic',
-                username: 'username',
-                password: 'password',
-              },
-            },
-          },
-          adyen: {
-            adyenMerchantAccount1: {
-              apiBaseUrl: 'apiBaseUrl',
-              apiKey: 'apiKey',
-              clientKey: 'clientKey',
-              legacyApiBaseUrl: 'legacyApiBaseUrl',
-            },
-          },
-          logLevel: 'DEBUG',
-          addCommercetoolsLineItems: false,
-        })
-        const config = await reloadModule('../../../src/config/config.js')
-        expect(
-          config.default.getModuleConfig().addCommercetoolsLineItems,
-        ).to.eql(false)
-      },
-    )
-
-    it(
-      'when addCommercetoolsLineItems is set as boolean true in config.js, ' +
-        'it should load as true value in module config',
-      async () => {
-        process.env.ADYEN_INTEGRATION_CONFIG = JSON.stringify({
-          commercetools: {
-            ctpProjectKey1: {
-              clientId: 'clientId',
-              clientSecret: 'clientSecret',
-              apiUrl: 'host',
-              authUrl: 'authUrl',
-              authentication: {
-                scheme: 'basic',
-                username: 'username',
-                password: 'password',
-              },
-            },
-          },
-          adyen: {
-            adyenMerchantAccount1: {
-              apiBaseUrl: 'apiBaseUrl',
-              apiKey: 'apiKey',
-              clientKey: 'clientKey',
-              legacyApiBaseUrl: 'legacyApiBaseUrl',
-            },
-          },
-          logLevel: 'DEBUG',
-          addCommercetoolsLineItems: true,
-        })
-        const config = await reloadModule('../../../src/config/config.js')
-        expect(
-          config.default.getModuleConfig().addCommercetoolsLineItems,
-        ).to.eql(true)
-      },
-    )
-
-    it(
-      'when addCommercetoolsLineItems is set as string false in config.js, ' +
-        'it should load as false value in module config',
-      async () => {
-        process.env.ADYEN_INTEGRATION_CONFIG = JSON.stringify({
-          commercetools: {
-            ctpProjectKey1: {
-              clientId: 'clientId',
-              clientSecret: 'clientSecret',
-              apiUrl: 'host',
-              authUrl: 'authUrl',
-            },
-          },
-          adyen: {
-            adyenMerchantAccount1: {
-              enableHmacSignature: 'false',
-            },
-          },
-          logLevel: 'DEBUG',
-          addCommercetoolsLineItems: 'false',
-        })
-        const config = await reloadModule('../../../src/config/config.js')
-        expect(
-          config.default.getModuleConfig().addCommercetoolsLineItems,
-        ).to.eql(false)
-      },
-    )
-
-    it(
-      'when addCommercetoolsLineItems is set as string true in config.js, ' +
-        'it should load as true value in module config',
-      async () => {
-        process.env.ADYEN_INTEGRATION_CONFIG = JSON.stringify({
-          commercetools: {
-            ctpProjectKey1: {
-              clientId: 'clientId',
-              clientSecret: 'clientSecret',
-              apiUrl: 'host',
-              authUrl: 'authUrl',
-            },
-          },
-          adyen: {
-            adyenMerchantAccount1: {
-              enableHmacSignature: 'false',
-            },
-          },
-          logLevel: 'DEBUG',
-          addCommercetoolsLineItems: 'true',
-        })
-        const config = await reloadModule('../../../src/config/config.js')
-        expect(
-          config.default.getModuleConfig().addCommercetoolsLineItems,
-        ).to.eql(true)
-      },
-    )
-  })
-
   describe('generateIdempotencyKey', () => {
     it(
       'when generateIdempotencyKey is set as boolean false in config.js, ' +
